+++ Firmware/MotorControl/controller.cpp
@@ -65,6 +65,18 @@ void Controller::start_anticogging_calibration() {
     }
 }
 
+// Slowly drive in the negative direction at homing_speed until the min endstop is pressed
+// When pressed, set the linear count to the offset (default 0), and then
+bool Controller::home_axis() {
+    if (axis_->min_endstop_.config_.enabled) {
+        set_vel_setpoint(-config_.homing_speed, 0.0f);
+        axis_->homing_state_ = HOMING_STATE_HOMING;
+    } else {
+        return false;
+    }
+    return true;
+}
+
 /*
  * This anti-cogging implementation iterates through each encoder position,
  * waits for zero velocity & position error,