+++ Firmware/MotorControl/odrive_main.h
@@ -122,6 +122,7 @@ inline ENUMTYPE operator ~ (ENUMTYPE a) { return static_cast<ENUMTYPE>(~static_c
 #include <controller.hpp>
 #include <motor.hpp>
 #include <trapTraj.hpp>
+#include <endstop.hpp>
 #include <axis.hpp>
 #include <communication/communication.h>